Add files via upload
This keylogger is capable of recording all the keyboard and mouse input! It will even record freelance of the language settings as a result of it logs the physical keys on the keyboard initial. Next, by victimisation Associate in Nursing whimsical keymap with human friendly names, it interprets the machine keys to one thing that we are able to perceive. It additionally possesses mail causing capabilities thus you'll simply schedule the logfile to be sent via mail, manually time input for each twelve hours. Additionally, to the current, it'll additionally keep the logfile encrypted. Therefore, I have prepared another program which is able to decode the logfile on execution.
 I trust in making this project(tool) just for learning.  This keylogger is not for any black or immoral functions, you risk losing plenty. This is often just for instructional functions. The keylogger that you just area unit about to see are going to be able to cause some serious harm, thus please use it in an exceedingly legal and accountable method.